Added parameters in SubsonicClient constructor, refactored Extension method to proper class.
